requirements.txt
beautifulsoup4
discord.py
Flask
python-dotenv
requests
Gunicorn